Return the memory after scrypt to the OS.
Previously a runtime.GC was being invoked which force it to release the
memory as far as the garbage collector is concerned, but the memory was
not released back to the OS immediatley.  This modification allows the
memory to be released immediately since it won't be needed again until the
next wallet unlock.